+++ app/Config/Auth.php
@@ -133,7 +133,7 @@ class Auth extends \Myth\Auth\Config\Auth
 	 *
 	 * @var bool
 	 */
-	public $allowRegistration = true;
+	public $allowRegistration = false;
 
 	/**
 	 * --------------------------------------------------------------------