+++ app/Config/Routes.php
@@ -46,6 +46,7 @@ $routes->group('/', ['filter' => 'login'], function($routes){
 		$routes->get('add', 'CDR\Users::add');
 		$routes->post('add', 'CDR\Users::store');
 		$routes->get('list', 'CDR\Users::list');
+		$routes->get('status', 'CDR\Users::doStatus');
 	});
 });